feat(mining): unify galgame captions into lookup popup, drop texthooker tab
- 抽出 GalgameSessionController（单例）：搬 TexthookerPage 的启动+注入、文本
  轮询、语音配对缓存、窗口绑定+封面逻辑；文本轮询改喂 DesktopLookupService
  .submitText（与剪贴板查词同去向，默认落悬浮查词面板）
- 删 texthooker 首页 tab + TexthookerPage（HomeTab.texthooker / homeActiveTabs
  分支 / nav item / buildTabContent / barrel / widget 测试）；WS 通路
  （TexthookerService + ws client/host + texthooker_enabled 设置）保留
- 游戏库点游戏 → GalgameSessionController.launch（不再 push 独立页）；games tab
  改由 Platform.isWindows 门控（galgame 注入本就 Windows-only，默认可见）
- 制卡富化接入点从被删页挪到查词制卡出口：HomeDictionaryPage.onMineEntry 覆写，
  有活跃 galgame 会话时带该句配对语音 + 窗口封面（GIF/PNG），否则 super（Never
  break：剪贴板/阅读器/有声书/全局查词制卡不变）
- i18n desktop_clipboard_destination 改名「查词弹窗位置」/ "Lookup popup position"
- desktop_clipboard_enabled 默认改 true（开箱即用）
