+++ hibiki/lib/i18n/strings_zh-CN.i18n.json
@@ -1780,7 +1780,7 @@
   "video_danmaku_manual_no_result": "未找到匹配的番剧。",
   "video_danmaku_manual_network_error": "网络错误，请检查连接后重试。",
   "video_danmaku_manual_server_error": "搜索失败，请稍后重试。",
-  "desktop_clipboard_destination": "剪贴板查词显示位置",
+  "desktop_clipboard_destination": "查词弹窗位置",
   "desktop_clipboard_destination_main": "主窗口",
   "desktop_clipboard_destination_panel": "悬浮面板",
   "desktop_clipboard_destination_transient": "光标处弹卡",
