+++ hibiki/lib/i18n/strings_zh-HK.i18n.json
@@ -1780,7 +1780,7 @@
   "video_danmaku_manual_no_result": "No matching anime found.",
   "video_danmaku_manual_network_error": "Network error. Check your connection and try again.",
   "video_danmaku_manual_server_error": "Search failed. Try again later.",
-  "desktop_clipboard_destination": "Clipboard lookup destination",
+  "desktop_clipboard_destination": "查詞彈窗位置",
   "desktop_clipboard_destination_main": "Main window",
   "desktop_clipboard_destination_panel": "Floating panel",
   "desktop_clipboard_destination_transient": "Popup at cursor",
