+++ hibiki/lib/main.dart
@@ -1444,7 +1444,7 @@ class _HoshiReaderAppState extends ConsumerState<HoshiReaderApp>
                       // a MacosWindowScope and can use native MacosScaffold/ToolBar.
                       // MacosTheme is derived from the SAME live ColorScheme as the
                       // rest of the app. The sidebar destinations come from the
-                      // dynamic HomeTab list (video/texthooker toggles) so they
+                      // dynamic HomeTab list (video/games toggles) so they
                       // stay in lock-step with HomePage's rail; selection is shared
                       // via homeShellTabNotifier. Hide the sidebar while a media
                       // item (reader/video) is open so reading is full-width; the
@@ -1469,10 +1469,10 @@ class _HoshiReaderAppState extends ConsumerState<HoshiReaderApp>
                                   : buildHibikiMacosSidebar(
                                       activeTabs: homeActiveTabs(
                                         // 「视频」tab 已毕业为常驻（原
-                                        // experimentalVideoEnabled 恒 true）。
+                                        // experimentalVideoEnabled 恒 true）。games
+                                        // （galgame 库）仅 Windows；macOS 根侧栏此处
+                                        // 恒 false（gamesEnabled 缺省），不显示。
                                         videoEnabled: true,
-                                        texthookerEnabled:
-                                            appModel.texthookerEnabled,
                                       ),
                                     ),
                               child: child!,
