+++ hibiki/test/pages/home_tab_games_test.dart
@@ -1,8 +1,9 @@
 import 'package:flutter_test/flutter_test.dart';
 import 'package:hibiki/src/pages/implementations/home_page.dart';
 
-/// 守卫「游戏」tab（galgame 库）在首页顶层导航中的可见性与位置。游戏库与 texthooker
-/// 同属 galgame 沉浸制卡，调用方以同一实验开关传入 [homeActiveTabs] 的 gamesEnabled。
+/// 守卫「游戏」tab（galgame 库）在首页顶层导航中的可见性与位置。galgame UX 统一后 games
+/// 是唯一的 galgame 入口（点游戏 → 台词进悬浮查词面板），生产里由 [homeActiveTabs] 的
+/// gamesEnabled = Platform.isWindows 门控（galgame 引擎-hook 注入本就 Windows-only）。
 void main() {
   test('HomeTab 枚举包含 games', () {
     expect(HomeTab.values, contains(HomeTab.games));
@@ -10,36 +11,30 @@ void main() {
 
   test('gamesEnabled 关闭（默认）时不出现', () {
     expect(
-      homeActiveTabs(videoEnabled: true, texthookerEnabled: true),
+      homeActiveTabs(videoEnabled: true),
       isNot(contains(HomeTab.games)),
     );
     expect(
-      homeActiveTabs(
-        videoEnabled: false,
-        texthookerEnabled: false,
-        gamesEnabled: false,
-      ),
+      homeActiveTabs(videoEnabled: false, gamesEnabled: false),
       isNot(contains(HomeTab.games)),
     );
   });
 
-  test('gamesEnabled 开启时紧邻 texthooker 之后、设置之前', () {
+  test('gamesEnabled 开启时夹在词典与设置之间', () {
     final List<HomeTab> tabs = homeActiveTabs(
       videoEnabled: true,
-      texthookerEnabled: true,
       gamesEnabled: true,
     );
-    final int texthooker = tabs.indexOf(HomeTab.texthooker);
+    final int dict = tabs.indexOf(HomeTab.dictionaries);
     final int games = tabs.indexOf(HomeTab.games);
     final int settings = tabs.indexOf(HomeTab.settings);
-    expect(games, equals(texthooker + 1));
+    expect(games, equals(dict + 1));
     expect(settings, equals(games + 1));
   });
 
-  test('texthooker 关但 games 开：games 落在词典与设置之间', () {
+  test('无视频时 games 仍夹在词典与设置之间', () {
     final List<HomeTab> tabs = homeActiveTabs(
       videoEnabled: false,
-      texthookerEnabled: false,
       gamesEnabled: true,
     );
     final int dict = tabs.indexOf(HomeTab.dictionaries);
