--- hibiki/test/pages/texthooker_page_test.dart
@@ -1,41 +1,0 @@
-import 'package:flutter/material.dart';
-import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'package:flutter_test/flutter_test.dart';
-import 'package:hibiki/src/pages/implementations/texthooker_page.dart';
-import 'package:hibiki/src/sync/texthooker_service.dart';
-
-void main() {
-  setUp(() => TexthookerService.instance.clear());
-  tearDown(() => TexthookerService.instance.clear());
-
-  testWidgets('renders incoming lines reactively', (WidgetTester tester) async {
-    await tester.pumpWidget(
-      const ProviderScope(child: MaterialApp(home: TexthookerPage())),
-    );
-    await tester.pump();
-
-    expect(find.text('第'), findsNothing);
-
-    TexthookerService.instance.appendLine('第一行');
-    await tester.pump();
-    // 分词后可能拆成多个 span，逐字降级时「第」是独立 span。
-    expect(find.textContaining('第'), findsWidgets);
-
-    TexthookerService.instance.appendLine('第二行');
-    await tester.pump();
-    expect(find.textContaining('二'), findsWidgets);
-  });
-
-  testWidgets('clear button empties the list', (WidgetTester tester) async {
-    TexthookerService.instance.appendLine('行X');
-    await tester.pumpWidget(
-      const ProviderScope(child: MaterialApp(home: TexthookerPage())),
-    );
-    await tester.pump();
-    expect(find.textContaining('行'), findsWidgets);
-
-    await tester.tap(find.byIcon(Icons.delete_outline));
-    await tester.pump();
-    expect(find.textContaining('行'), findsNothing);
-  });
-}
